ipc: ARG_TYPE_NONE passing non-zero value
When a function takes no argument the IPC bindings in config.h will use
ARG_TYPE_NONE as the argument type.

When this type is used a dummy Arg argument is used to pass a value of 0
to the function (like keybindings do).

The IPC patch would set a float value to 0 which had the side effect of
potentially crashing dwm if the function were to optionally also accept
a client reference to be passed in via the argument.

In C a float is represented as 4 bytes, whereas the IPC patch changes the
Arg i (int) value to a long which uses 8 bytes. The Arg is a union struct
which means that the different data types are stored in the same memory
location.

When memory for an Arg instance is reserved 8 bytes will be
allocated. Setting a float value to 0 most likely only clears the first
4 bytes (32 bits) of that memory location, leaving remnant memory data
in the remaining 4 bytes. Ultimately this means that the Arg value may
have a random value rather than 0.

This was fixed by setting an int (now long) to 0, which clears all 8 bytes.