Removing pinned workspace color scheme.
The reason is that although it was a nice idea it clashes with the
other schemes for the bar, making it counter-intuitive to for the
user to predict what color is being used and when.

This particularly conflicts with the occupied color scheme, and one
will always take precedence over the other. Maybe the occupied color
scheme should also be removed.