Several changes (AutoReduceMaster, KeyPressRelease, status2d, and more)
AutoReduceMaster
----------------
Introduced a new feature where if you have multiple master windows
and you close one of them, then the remaining master clients will
occupy the remaining space rather than a stack window taking it's
place. Essentially this means that if nmaster > 1 and the client
being removed is a master client, then nmaster is reduced by one.
Fixed a bug when checking whether a functionality is enabled or
not and the functionality is a long value > int.

KeyPressRelease
---------------
Included the keypressrelease patch to experiment with better
options for the holdbar patch and proposed implementation of the
latter.

statusbutton
------------
Removed statusbutton in favour of just another status.

status2d
--------
Generalised the status2d functions to be drw_2dtext and enabled
status2d markup for both tags and the layout symbol. This adds
some flexibility when it comes to tag colors and indicators.