Removing feature to ignore configure requests while the window is being moved or resized.
The reason for the feature in the first place was Alacritty awkwardly changing
size when it was being moved from one monitor to another.

The side effects of ignoring these configure requests was that the content of
the terminal window would be cut off as Alacritty would change size internally
anyway - often forcing you to manually resize the window just a bit for it to
display correctly again. This was more annoying than the original issue hence
rolling this change back.