Fix memory leak in dwm relating to updating the bar
The XSync call here seemingly causes a memory leak when a series
of XCopyArea calls are made without calling XNextEvent. This info
comes from a 20-year-old post:

ref. https://root.cern.ch/root/roottalk/roottalk01/1040.html

The leak can be seen if spamming IPC commands that trigger the bar
to be updated while monitoring the process' memory usage.

On the same note continuously triggering a function that does
nothing will not increase memory usage.

The XSync call flushes the output buffer and then waits until all
requests have been received and processed by the X server.

ref. https://tronche.com/gui/x/xlib/event-handling/XSync.html

If we pass True instead of False to XSync then that also will result
in no memory leak, but at the same time we would be losing updates.

Removing the XSync call altogether avoids the memory leak and so far
no side effects of this has been observed with regards to updating
the bar, which begs the question of why the call was there to begin
with.

This was not an issue in dwm where actions were restricted to be
triggered via the event loop.

On the off chance that the flushing of the output buffer was needed
for whatever reason we are here replacing the XSync call with XFlush.