xresources: refactoring for foreground colours support
Prior to this all the flexwintitle colour schemes relied on the
titlenormfgcolor and titleselfgcolor values for the font / foreground
colours for window titles in the bar.

This worked fine for the built-in theme that comes with dusk, but
was a hindrance when it comes to custom colour schemes via xresources
because the one font colour would have to look OK for all arrangement
options - which would ultimately force all the flexwintitle colours
to all be dark or all be light.

The refactoring here removes the dependency on the variable names in
config.h by referring to the colors array directly, which in principle
should prevent these changes from breaking existing configurations.