sortscreens: allow users to override the RIGHTOF macro
By default the sortscreens logic is such that it is based on the
y-axis unless all the monitors are at the same y-axis, in which
case the monitors are ordered based on the x-axis.

This change allows the user to meddle with this by adding their own RIGHTOF
function in the config.h file, e.g. to only take the x-axis into account.

Or for example one could order the screens from right to left.