powerline: nitpicks for powerlines at the start and end of the bar
The previous iteration simply got rid of the powerlines if they
were found to be placed at the start or end of the bar. This was
because if drawn they it would look weird having a slash or arrow
that comes from nothing or goes to nothing, and removing them gave
a clear sharp edge.

This resulted in some annoyance, however, given that the workspace
module partially depends on it, and it would cause the leftmost
workspace icon to appear smaller than the rest.

I refactored this into:
   - not dropping the powerlines if they are at the bar edges and
   - making sure that they render as a solid block and
   - reducing the size of the powerlines by half

This gives to my eyes a more appropriate padding for the first and
last bar module when cornered by a powerline.