applyrules: _KDE_NET_WM_WINDOW_TYPE_OVERRIDE
Apparently some nasty applications that declare the window type of
_KDE_NET_WM_WINDOW_TYPE_OVERRIDE expect to not be managed by the
window manager. More so an application can declare more than one
window type in the _NET_WM_WINDOW_TYPE atom, and the aforementioned
override may not be the first window type in the list.

The changes here introduces:
   - a rule in the default configuration filtering on a window type
     of _KDE_NET_WM_WINDOW_TYPE_OVERRIDE and making such windows
     unmanaged
   - logic in the applyrules function that allows for all window
     types listed in the _NET_WM_WINDOW_TYPE atom to be checked
     for rule matching purposes

An example application relying on the _KDE_NET_WM_WINDOW_TYPE_OVERRIDE
window type is the Nextcloud client applet, which is a small window
that is supposed to show up near the systray icon when pressed.