unmanage performance refactoring
This involves performance optimisations when it comes to killing
multiple windows in one go.

Let's say that for testing purpopses that you open 100 terminal
windows, then do a Super+a to select all, unselect one, then do a
Super+q to kill all the selected windows.

When this happens client messages are sent to each window asking
for them to be deleted / closed. This is then followed by a surge
of UnmapNotify events coming from the X server. Handling of each
such event results in unmapnotify and subsequently the unmanage
function being called.

Prior to this commit each call to unmanage would trigger a focus,
an arrange, and a drawbar call. The effect of that is a high CPU
spike and noticeable slow motion where the remaining window is
being resized 99 times depending on the layout used. In practice
this whole scenario may take only a second to complete but the
behaviour can seem a bit jarring nonetheless.

With this commit the unmanage function no longer handles focus
change, arrangement or updating of the bar. Instead all
UnmapNotify events will be handled in one go and once all events
have been processed the focus, arrange and drawbar logic takes
place. This happens in a new function structurenotify that handles
UnmapNotify, DestroyNotify and ConfigureNotify events.

The reason why it handles more than one event type has to do with
how all of these event types fall under the same event masks of
StructureNotifyMask and SubstructureNotifyMask. The following web
page gives a good overview of event masks vs event types:

https://tronche.com/gui/x/xlib/events/processing-overview.html