scratchpads: edge case involving multiple scratchpads
There was an issue such that if multiple windows are
assigned to a scratchpad and these are brought into view
on an empty workspace, then when hiding the windows only
one of them would actually be hidden.

This has to do with how the original showhide was removed
in favour of having the arrange function control the
behaviour combined with that no arrange takes place if
there are no clients to arrange.

The showwsclients calls here are redundant, but were added
for consistency with the hide call which is the actual fix
for this issue.