autostart: kill processes after systray icons have been unmapped
The reason for this is that if the autostart array contains a systray
program like nm-applet then there will be a race condition of sorts
where the nm-applet process is killed which ultimately means that it
will take the systray window (icon) with it as well.

Prior to this change the code would proceed to clean up and unmap all
systray icons. Later when the XSync(dpy, False); call was made the X
server would hit a segmentation fault, presumably due to how the
systray window is being torn down and unmapped at the same time.

We are addressing this issue by cleaning up and unmapping all systray
icons (and syncing up with the X server) before proceeding to kill
processes.

Addresses issue #14.