restart: restrict the number of workspace properties that can be read
After restarting do not read more than _NET_NUMBER_OF_DESKTOPS properties
from the _DUSK_WORKSPACE property.

Reasoning:
If the property does not exist then XGetWindowProperty simply returns as
a non-success in retrieving the property.

If the number of properties is, say, 9 and you try to read the 10th
property using XGetWindowProperty then it will return a BadValue error
crashing X as well as the window manager.

Previously we read properties based on the number of workspace rules in
the config. This had the side effect that if you were to increase the
number of workspace rules and then restart then it would crash with a
BadValue error.

The _NET_NUMBER_OF_DESKTOPS atom represents the number of workspaces of
the current (and thus also the previous) session so we can deduce how
many properties the previous session will have written when restarting.

An alternative approach might have been to put a cap on the number of
workspaces and e.g. always writing 64 values to this property.