AlwaysOnTop - tentative change to align with wiki documentation
The AlwaysOnTop flag only had an effect if the FocusedOnTop
functionality was enabled. I.e. having that flag set would
not result in a window remaining above others.

This is a tentative change to allow the AlwaysOnTop flag to
have an effect when the FocusedOnTop functionality is disabled.
The approach here may change in the future if a better solution
is found.