wspreview: refactoring setup and teardown logic
This is to address an issue where if the resolution of the
monitor was changed then the size of the preview windows would
remain relative to the previous (original) resolution rather
than being pfact (default 1/4th) of the current monitor
resolution.

Additionally move common code into separate functions to avoid
duplication.