ViewOnWs: move functionality check outside of movetows function
In principle this should just be refactoring and cause no functional
changes.

These changes were introduced due to the hacky way sendtowsbyname was
implemented merely toggling the ViewOnWs off and on again.

In principle the functionality toggle could be made redundant by
offering two functions - one that merely moves clients to another
workspace and the other which additionally changes the view to the
workspace the clients were sent to.

Currently we have sendtowsbyname which only moves a client to another
workspace and movetowsbyname which additionally follows the view to
the destination workspace. Word-wise "move" doesn't clarify that the
view follows the client(s) to the destination workspace.

Potentially it may be worth changing function names to be more
intuitive.

For example:
   - sendtowsbyname vs
   - viewonwsbyname
   - sendalltowsbyname vs
   - viewallonwsbyname
   - moveallfromwsbyname (moves all to selected workspace)
   - sendwsdir vs
   - viewonwsdir (not to be confused with existing viewwsdir)