incnmaster: making decrementing the number of master clients more intuitive
Prior to this change the nmaster could be increased indefinitely by
continuously hitting the increment keybinding. When decrementing the
nmaster there is no visible feedback making it appear as nothing has
changed. With this change we make nmaster increments and decrements
relative to the number of tiled clients if we have less tiled clients
than nmaster.

Additionally this change has refactoring due to adding the numtiled
helper function.