Refactor atom initialisation
This replaces XInternAtom(3) with XInternAtoms(3) which allows for
multiple atoms to be retrieved in a single protocol call rather than
having one call for each atom.

The enums and lists have also been moved to ewmh.h which should make
it easier to add new atoms when needed as everything is in one place.