configurerequest: respect requests to move a window to absolute x and y positions
The window manager may receive a configure request event for a window which in
practice means that the desire is that the window is to move to a given position
and/or change size.

Most window managers and desktop environments accepts a given position as being
absolute in the space that the monitors occupy.

dusk is based on dwm and dwm treats windows (clients) as belonging to a monitor
and as such when a configure request comes in the position is always going to be
treated as being relative to the monitor the client is on.

I.e. consider the following command:

   xdotool getactivewindow windowmove 100 100

and we have a multi-monitor setup with a vertical screen on the left and a
horizontal screen on the right.

The left screen's top left corner will have position 0x0 while the right screen
may have top left corner position of 1024x600.

When the command is executed in dwm in a terminal on the right hand monitor the
window will be placed at position 1124x700 (i.e. the position is relative to the
monitor the window is on). In most other window managers the terminal would have
moved to position 100x100 as requested, which would be at the top of the left
hand monitor.

While the relative positioning of a dumb terminal works well in this case many
programs that send configurerequests typically also have some awareness of the
position of the current window. This feature then has some unintended side
effects where the window ends up jumping around the screen because it is
trying to place itself relative to its current position - which is not where it
asked to be placed previously.

These changes makes dusk behave the the same way as most other window managers
in that it will, by default, treat configure request positions as being
absolute. That should fix most programs that typically misbehave in dwm and the
use of flags like IgnoreCfgReqPos to disregard a window's request to change
position should in principle be redundant.

For those cases where one may actually want configure requests to be restricted
to be relative to the current monitor a new flag CfgReqPosRelativeToMonitor was
introduced which will enable the old behaviour on a per application basis as
per client rules.