Hide windows above rather than all to the left
In dwm when windows are hidden they are merely moved to a negative
x position and are moved back in when shown.

This change hides windows at a negative y position instead. In
practice this should make no difference to normal behaviour when it
comes to changing workspaces. When using a compositor that handles
animation of windows then it may be nicer to have fly up and come
trickling down rather than having them slide off to the left and
back.

Another approach could be to unmap and re-map windows when they are
shown and hidden.