Use window mapping when hiding and showing windows
rather than moving the window to a negative position and back.

This surprisingly results in a performance boost as tested with
the following script that switches between workspace 7 and 8
repeatedly.

for I in $(seq 1000); do
	duskc --ignore-reply run_command viewwsbyname 7
	duskc --ignore-reply run_command viewwsbyname 8
done

Workspaces 7 and 8 having 12 and 7 windows respectively. In total
2000 workspace switches are performed.

Prior to this change we got an average runtime of 50.653 seconds.

$ time ./time_ws_change.sh
./time_ws_change.sh  2.91s user 5.11s system 15% cpu 52.511 total

With this change we got an average runtime of 41.324 seconds.

It also ran faster without a compositor.

This suggests a performance improvement of about 21.31% when changing
views (this despite the need to handle additional UnmapNotify events).