always on top: refactoring and improving functionality
This largely involves:
   - moving XRaiseWindow and AOT logic into a new function raiseclient and
   - make restack call raiseclient instead of XRaiseWindow and
   - removing AOT logic from the focus function
   - plus replacing other XRaiseWindow calls with a call to raiseclient

Additionally we remove the restack calls from movemouse and resizemouse as
these were essentially redundant as the logic in buttonpress will call focus
and restack for a client before changing the click action to ClkClientWin.

Granted a restack may be missed if movemouse or resizemouse is used via IPC
commands. Will revisit if that does ever turn out to be an issue.

Clients that are made fullscreen are still allowed to be placed above clients
that are marked as being always on top.