Rename restartsig variable restartwm
As per the original restartsig patch the variable was named just restart
which makes more sense. When creating the restart function to more clearly
differentiate between restarting and quitting the window manager the
variable name was changed to restartsig.

The variable can be set in other ways than only via a kill -HUP signal so
the name can be misleading. Decided to rename this restartwm instead to
make this more clear.