Introduce throttling of property notifications.
Some programs may end up spamming property notifications rendering the window
manager unable to do anything else than to process the backlog of these for as
long as it takes. One may think of it as a DOS attack for the window manager.

Example symptoms include:
   - the dusk process running at 100%
   - the bar stops responding
   - the bar becomes black
   - the mouse can still be moved, but focus can't be changed between windows
   - i.e. X is running fine
   - application windows are working / updating fine
   - duskc will not work (may crash with a segmentation fault)
   - memory usage for the dusk process is significantly increased
   - a kill -HUP on the dusk process makes no difference
   - if left without intervention the window manager fully recovers when all
     the property notifications have been checked and processed (assuming
     the origin of these stops sending them)

This change introduces throttling of such property notification messages.
If a notification for the same state and atom is received more than three
times in a row for the same window then the window manager will start ignoring
property notifications and explicitly purge pending notifications.