clientmessage: message to remove maximized state should only restore float position when client is floating
This was found with Nautlus which sends client messages to remove fullscreen and to remove
maximized state on startup.

clientmessage: received message type of _NET_WM_STATE (353) for client Loading…
    - data 0 = _NET_WM_STATE_REMOVE (0)
    - data 1 = _NET_WM_STATE_MAXIMIZED_VERT (354)
    - data 2 = _NET_WM_STATE_MAXIMIZED_HORZ (355)

The side effect of this was that the program would start as tiled then being moved to another
location while still being tiled. As soon as any other change happens then the window will
snap back to it's designated position.

With this change that issue should no longer occur.