unmanaged: allow unmanaged windows have the possibility of grabbing input focus
By default dusk would let the focused window keep input focus even
if the focus may drift to other windows that are not managed by the
window manager. This in turn prevents such windows from grabbing
input focus.

This change makes sure to revert input focus back to the root window
in the event that a LeaveNotify event is received.