KeepInputFocusOnLeaveNotify: adding "firefox" flag
firefox will hide its hamburger menu as soon as the main window loses
input focus, and the main window will lose input focus as soon as the
user moves the mouse cursor to the hamburger menu.

Ultimately this means that firefox does not check whether focus would
reside on the menu itself before removing it.

The KeepInputFocusOnLeaveNotify flag is essentially a hack specifically
for firefox such that the program will retain input focus after having
received a LeaveNotify event for that window.