xrdb: remove "dusk." prefix from bespoke resource strings ref. #16
By default bespoke resources such as these:

 /* Xresources preferences to load at startup. */
 static const ResourcePref resources[] = {
 	{ "dmenu.norm.fg.color", STRING, &dmenunormfgcolor },
 	{ "dmenu.norm.bg.color", STRING, &dmenunormbgcolor },
 	{ "dmenu.sel.fg.color", STRING, &dmenuselfgcolor },
 	{ "dmenu.sel.bg.color", STRING, &dmenuselbgcolor },
 	{ "dmenu.font", STRING, &dmenufont },
 };

would be prefixed with "dusk." when looking up said resource strings.

I.e. "dusk.dmenu.sel.bg.color" would be looked up for the selected
background colour.

This was not intuitive and it makes more sense to add "dusk." to the
resource string defined in the above config if desirable as such.