scratchpads: only hide single scratchpads when they have focus
The previous behaviour was that:
   - if one or more scratchpads is visible and the user hits the
     keybinding for the scratchpad then the windows will be hidden

The new behaviour is that:
   - if a single scratchpad is visible and has focus and the user
     hits the keybinding for the scratchpad then the window will
     be hidden
   - if a single scratchpad is visible, but does not have focus,
     and the user hitst he keybinding for the scratchpad then the
     window will receive focus
   - if there are more than one window associated with a scratchkey
     then all windows will show or hide as with the previous
     behaviour
   - the same goes if there is a single scratchpad window that is
     on another workspace (provided that it is not moved to the
     current workspace)

What this is trying to fix, or to make more user friendly, is when
there are multiple floating windows on the screen and a scratchpad
window may end up behind other windows. Typically in this scenario
one may expect that the scratchpad should come into view (on top)
when hitting the scratchpad shortcut, only to find that the window
hides and one has to hit the shortcut a second time in order to
see the window.

Naturally this has a side effect of having to hit the keyboard
shortcut twice to hide a scratchpad if it does not already have
focus, but hopefully this should be more intuitive than having to
hit the shortcut twice to give a visible scratchpad window focus.

If the new behaviour turns out to be detrimental to other people's
expectations then I will add a configuration toggle for this.