Allow previous workspace to be stored when using the mouse
Prior to this commit the previous workspace would only be stored when
using keybindings to change workspaces. This would allow for workspaces
to be viewed quickly by doing e.g. Super+5+5 to glance at workspace 5
and go back to the previous workspace by hitting 5 again without letting
go of the modifier key.

When changing workspaces by clicking on the workspace icon on the bar
using the mouse this the previous workspace would not be stored. This
was because it would be confusing if clicking a second time on an icon
(or double clicking for that matter) would lead you back to another
workspace.

There was an edge case though where if the user were to select a
workspace with the mouse, and were to use the keybinding to move to the
same workspace - then the expectation would be to move to the previous
workspace and this was not the case.

While it is a peculiar situation this issue would become more apparent
if one were to set up a dedicated keybinding that moves back to the
previous workspace.

This change makes sure that the previous workspace is recorded both
when changing workspaces using keybindings and when using the mouse.