swallow: addressing issues related to stack order when unswallowing client
When a GUI application, e.g. nautilus, swallows a tiled terminal, and that
GUI application is made floating before closing (or unswallow is otherwise
triggered), then the terminal application should also be floating at the
same size as the GUI application (as far as size restrictions allow).

A problem with this was that a tiled window is placed below the bar so when
the GUI application was closed then the terminal application would be
floating and in the same position - but it would disappear behind the tiled
windows as they would have had focus in the meantime and thus appear higher
up in the stack order.

To address this we are now:
   - allowing the stacking order to be inherited in replaceclient and
   - if the old client was the selected client then allow the new client
     to become the selected client