enternotify: favour client under the mouse cursor
When moving the mouse cursor quickly across multiple windows, for
example if there are several overlapping floating windows, then this
will generate enternotify events in the X server.

Given that the focus function may call skipfocusevents in order to
not process any subsequent enternotify events this can lead to a
race condition where the mouse cursor has already reached its
destination window, but focus goes to another window prior to that
on the path to the desired window.

In this context it makes sense to favour the window under the mouse
cursor at the time when the event is processed - and fall back to
the window the enternotify event is for should the mouse cursor not
be on top of a client window.

In principle there could still be room for race conditions depending
on how far the mouse moves while all of this is checked, so it makes
sense removing the call to skip focus events in the focus function.