hotplugging: at least one visible workspace per monitor
Depending on workspace rules it was found that with three or more monitors
it was possible to end up with one screen with two visible workspaces and
another screen with workspaces assigned - but none that were visible.

As such we introduce an effort to ensure that at least one of the
workspaces assigned to a monitor is visible.

Additionally it was decided to ensure that at most one workspace assigned
to a monitor is visible after hotplugging. This because it is a worse
experience if you only view one workspace at a time and unintuitively
end up with two visible workspaces after hotplugging, than it is to view
two workspaces and end up with only one visible workspace after
hotplugging.

These two changes also add some consistency when it comes to visible
workspaces after hotplugging a monitor.