storewsmask: refactoring logic as to when workspace mask is stored
It makes sense to always store the workspace mask whenever the view
changes. This so that when using e.g. focusurgent or movetowsbyname
then hitting MOD+<workspace you are on> will consistently bring you
to the previous workspace that was active before the action was
taken.

Also split logic into storewsmask and togglewsmask rather having
the function do two different things.