Make default workspace rules consistent across all workspaces.
Got feedback that perhaps the workspace rules should have consistent tiling / layout
in the default configuration and I agree. Having different layouts set or different
mfact was there as example configuration, but it makes more sense to have these
consistent in the default configuration and the user can manipulate this later if
they want to.

The workspace rules listed on the wiki has different layouts and mfact so that
should be good enough as an example configuration.

https://github.com/bakkeby/dusk/wiki/Workspaces