Ungrab pointer when spawning commands.
When spawning newer versions of xmenu using button clicks on the bar or otherwise
via the window manager the graphical window would simply not appear (instead exiting
quietly).

The reason for this was that xmenu tries to grab the mouse pointer which fails when
spawned via dusk.

Further analysis found that the mouse pointer could not be grabbed because it had
already been grabbed. It is not entirely clear how the mouse cursor is grabbed in
this case considering that XUngrabPointer is always called after moving or resizing
a window using the mouse for example, but there could potentially be something
obscure in relation to buttons being grabbed or underlying libraries.

In any case explicitly ungrabbing the mouse pointer before spawning windows seems
to do the trick.