placemouse: don't snap / arrange placement window while deciding the position of that window
When using placemouse and moving a window around there is a point when the client window
being moved will take the place of the client that it is hovering; when the swap takes place
an arrange is triggered which places all windows again. If doing this slowly it is noticeable
that the window being moved flickers into place and back again.

This change locks the client so that it is not affected by the arrange, i.e. it will not be
moved into the alotted tile while the placemouse is in action.