fullscreen: separating fullscreen and floating states
Inherited from dwm logic where when a window goes into fullscreen it is also marked as being floating.

This serves two purposes in dwm:
   - prevent the window from being tiled and
   - ensure that the window remains on top of tiled windows (placed
     above the bar window) following a restack

Linking the two (practically unrelated) states like this made it cumbersome to
differentiate between whether a window was floating or not in the context of
window swallowing. For example a tiled window could be in fullscreen, then being
swallowed by another window. When the window is later unswallowed it makes it
complicated to reason about whether the window was floating or not, considering
that it was in fullscreen and the other window may or not be in fullscreen.

It made more sense to separate the two states such that going into fullscreen
does not affect the floating state at all.