Revert input focus back to selected client on destroy notify
Some windows that are not managed by the window manager may grab input focus.
In some cases this was not given back when the unmanaged window closes, which could
lead to a window that has focus as far as the window manager is aware, but does not
have input focus.

Introducing a change to give the input focus back to the selected client when a
DestroyNotify event is received for an unmanaged window and the selected client
does not have input focus.

Also refactoring destroynotify to use if return rather than multiple if elseif elseif.