SystrayNoAlpha + xrdb compatibility issue
When changing colour scheme during runtime using xrdb the systray and
icons would sometimes keep the original colours.

This has to do with that changing the window attributes for systray
icons do not cause the window content to be redrawn; it just sets the
attributes for the next time the window content is drawn.

As such we need a way to force the application to redraw the content
of the systray icon. Calling XClearWindow for each systray icon does
not trigger a redraw unfortunately; it just makes the systray icons
disappear until the application decides to redraw the icon window
content.

One solution would be to send an Expose event for each window which
should in principle trigger a redraw from the application side.
Technically it is the X server that sends Expose events, but in
theory one could potentially have the window manager send those as
well.

Turns out that if we move the window out of the drawable area (i.e.
to a negative position) and back again then the X server will send
an Expose event for the window causing it to be redrawn. This change
makes sure to move the systray window to a negative position as part
of the xrdb call, and we leave the normal drawbar to bring them back
into view.

This issue is isolated to using the systray without alpha.