togglews: don't store previous workspace mask when using combo
Triggering togglews should bring you back to the previously
selected workspace(s) on the current monitor.

If you were to go to workspace 8 and then select workspaces
4, 5 and 6 using combo view, then when triggering togglews
it would bring you to workspace 4 and 5 as those were the
last shown workspaces as per combo logic.

This change ensures that the previous workspace mask is not
stored when combo is enabled allowing togglews to flip
between workspaces 8 and 4, 5 and 6 in this example when
triggered.