client rules: prevent assignment of unavailable workspaces
If config workspaces_per_mon is enabled then workspaces that are pinned to monitors that are
not available will be moved to a "dummy" monitor.

Workspaces that are assigned to such a monitor are not intended to be viewed; they are not
available and they are staying there until the surplus monitor is available.

That said there may be obscure edge cases where a client can get assigned to a workspace on
the dummy monitor or such a workspace being otherwise brought into view. If so we allow that
workspace to be brought to the selected monitor.

If a client rule is set to move a window to a workspace that is resting on the "dummy" monitor
then this will be ignored and it will fall back to assigning the window to the selected monitor.

This is because in this situation if the rule does not also include the SwitchWorkspace flag
then the window would simply disappear and the end user would not have any reasonable way to
find it.